as_root(): accept all KISS_SU values by default even if the value is not supported
This change ensures that niche tools that are not packaged in repositories
can be used without the need of it being implemented inside the package
manager itself. We don't support these tools, however most of them follow
a similar syntax as with 'sudo', and they can be run.

See: https://github.com/kiss-community/kiss/issues/12#issuecomment-799055508