kiss: make the install prompt only appear when extra dependencies
The prompt used to appear when more than 1 packages were to be installed, but
this also happened when they all were specified on the commandline (even when
no extra dependencies were resolved). This is kind of redundant behaviour.

Now it will only appear when the number of packages explicitly supplied to
it and number of packages after dependency resolution are unequal (because
dependencies were found on resolution). I assume this was meant to be the
original intent for the behaviour but was left behind even after pkg_install()
gained the ability to receive multiple arguments.

Signed-off-by: FriendlyNeighborhoodShane <shane.880088.supw@gmail.com>