kiss: add multilib support
Related GitHub issue: https://github.com/kiss-community/kiss/issues/20

Using readelf for fixing dependencies outputs only the soname of dynamic
libraries, which causes confusion between folders such as `/lib` and
`/lib32`.

This change ensures that we retrieve the correct path for the library
dependency by also using ldd. readelf provides the left handside of the
ldd dependency output, and we retrieve the right handside from that
information.

However, we don't want to call the dynamic loader lots of times, as that
would slow the operation. Therefore, we save ldd output in a buffer at
the beginning for each file, and we call that buffer instead when we
need to reference it.