Fix random dependency issue (#5)
Related GitHub issue: https://github.com/kiss-community/kiss/issues/5

kiss(1) will sometimes not install certain dependencies. This is caused
by a loop which installs dependencies from the kiss binary cache,
instead of rebuilding. The loop iterates the positional parameters
(packages to be build) and shifts the positional parameters for every
package installed from the binary cache. The mistake is, that the first
positional parameter being shifted away, might not be the same package
as the one of the current loop iteration. For example: If the packages
"dep2", "dep1" and "foopackage" are to be build, while only "dep1" has a
cached binary, the second loop iteration will shift once, falsely
removing "dep2" from the packages which still need to be build.

This patch uses a separate variable, to properly keep track of the
remaining packages to be build.